requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
auth0-python==4.8.0
Authlib==1.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.8
frozenlist==1.5.0
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
libcst==1.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
starlette-graphene3==0.6.0
strawberry-graphql==0.258.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
yarl==1.18.3
